fix delete methods

+++ src/main/java/com/example/ToDoList/service/implService/TaskServiceImpl.java
@@ -6,17 +6,23 @@ import com.example.ToDoList.exception.TaskNotFoundException;
 import com.example.ToDoList.model.entity.task.TaskEntity;
 import com.example.ToDoList.model.entity.user.UserEntity;
 import com.example.ToDoList.repository.TaskRepository;
+import com.example.ToDoList.repository.UserRepository;
 import com.example.ToDoList.service.TaskService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository repository;
 
+    private final UserRepository userRepository;
+
     @Override
+    @Transactional(readOnly = true)
     public TaskResponseDto findById(Integer id) {
         TaskEntity task = repository.findById(id)
                 .orElseThrow(() -> new TaskNotFoundException("Task with id - " + id + " not found"));
@@ -38,6 +44,7 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public TaskResponseDto finsByOwner(UserEntity owner) {
         TaskEntity task = repository.findByOwner(owner)
                 .orElseThrow(() -> new TaskNotFoundException("Task with owner - " + owner + " not found"));
@@ -58,17 +65,23 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public void deleteById(Integer id) {
+        if (!repository.existsById(Long.valueOf(id))) {
+            throw new TaskNotFoundException("Task with id " + id + " not found");
+        }
         repository.deleteById(id);
     }
 
     @Override
     public void deleteByOwner(UserEntity owner) {
+        if (!repository.existsByOwner(owner)) {
+            throw new TaskNotFoundException("Task for owner " + owner.getId() + " not found");
+        }
         repository.deleteByOwner(owner);
     }
 
     @Override
     public void create(TaskEntity entity) {
-        //entity.setOwner(userRepository.getById(4));
+        entity.setOwner(userRepository.getById(1));
         repository.save(entity);
     }
